replace assembly entry point with built in Rust
The existing model of building a binary as a library and then linking
that with assembly to make the actual binary is not conducive going
forward. This moves the assembly code into a Rust macro and allows the
user to use a macro to declare an entry point which is then wrapped
appropriately to build a functioning Xen binary. This change allows code
to be compiled on any platform that has a functioning Rust compiler.