Make xen bindings independent from standard library
When generating bindings, C types such as "int" are replaced with
"std::c_int" in the resulting bindings.  This representation
automatically brings in a dependency on Rust's standard library,
something that is not compatible with the #![no_std] and #![no_main]
directive typically found in bare metal applications.

This patch makes xen bindings independent of the standard library by
defining the size of the types.  For example, "std::c_int" above becomes
"xen_bindings_x86_64::c_int", which is defined in bindings_x86_64.rs.

That way baremetal applications can use the bindings without needing to
include the standard library.

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>