xen-ioctls: add features for XEN_DOMCTL_INTERFACE_VERSION
Allow downstream user of crate to select what interface version they
want by using Cargo features.

The default if no feature is selected is `0x16`.

To select the feature when xen-ioctls is a transitive dependency, simply
add it as a direct dependency with the feature enabled, and it will be
selected for all transitive dependencies as well.

Signed-off-by: Manos Pitsidianakis <manos.pitsidianakis@linaro.org>