xen-sys: add SAFETY docs to layout test unsafe{}s
The default CI pipeline for rust-vmm unconditionally errors for
undocumented unsafe blocks, so workaround the problem for now by adding
them manually in bindgen-generated layout tests.

The proper solution would be to use a custom pipeline for this
repository, or regenerate the bindings with a more recent bindgen
version which does not generate tests like this.

Signed-off-by: Manos Pitsidianakis <manos.pitsidianakis@linaro.org>